+++ core/src/main/java/io/github/therealmone/fireres/core/common/report/Report.java
@@ -1,0 +1,4 @@
+package io.github.therealmone.fireres.core.common.report;
+
+public interface Report {
+}
