+++ core/src/main/java/io/github/therealmone/fireres/core/common/report/ReportBuilder.java
@@ -1,0 +1,10 @@
+package io.github.therealmone.fireres.core.common.report;
+
+
+import io.github.therealmone.fireres.core.common.config.GenerationProperties;
+
+public interface ReportBuilder<T extends Report> {
+
+    T build(GenerationProperties properties);
+
+}
