--- core/src/main/java/io/github/therealmone/fireres/core/firemode/model/FireMode.java
@@ -1,22 +1,0 @@
-package io.github.therealmone.fireres.core.firemode.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class FireMode {
-
-    private StandardTemperature standardTemperature;
-    private MinAllowedTemperature minAllowedTemperature;
-    private MaxAllowedTemperature maxAllowedTemperature;
-    private FurnaceTemperature furnaceTemperature;
-    private List<FireModeSample> samples;
-
-}
