+++ core/src/main/java/io/github/therealmone/fireres/core/pressure/generator/ExcessPressureGenerator.java
@@ -2,6 +2,8 @@ package io.github.therealmone.fireres.core.pressure.generator;
 
 import io.github.therealmone.fireres.core.common.generator.PointSequenceGenerator;
 import io.github.therealmone.fireres.core.common.model.DoublePoint;
+import io.github.therealmone.fireres.core.pressure.model.MaxAllowedPressure;
+import io.github.therealmone.fireres.core.pressure.model.MinAllowedPressure;
 import io.github.therealmone.fireres.core.pressure.model.SamplePressure;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
@@ -13,13 +15,21 @@ import static io.github.therealmone.fireres.core.utils.RandomUtils.generateValue
 
 @RequiredArgsConstructor
 public class ExcessPressureGenerator implements PointSequenceGenerator<SamplePressure> {
+
     private final Integer time;
-    private final Double delta;
+
+    private final MinAllowedPressure minAllowedPressure;
+    private final MaxAllowedPressure maxAllowedPressure;
 
     @Override
     public SamplePressure generate() {
         val points = IntStream.range(0, time)
-                .mapToObj(i -> new DoublePoint(i, generateValueInInterval(-delta, delta)))
+                .mapToObj(i -> {
+                    val min = minAllowedPressure.getPoint(i).getValue();
+                    val max = maxAllowedPressure.getPoint(i).getValue();
+
+                    return new DoublePoint(i, generateValueInInterval(min, max));
+                })
                 .collect(Collectors.toList());
 
         return new SamplePressure(points);
