--- core/src/test/java/io/github/therealmone/fireres/core/generator/firemode/ExcessPressureGeneratorTest.java
@@ -1,33 +1,0 @@
-package io.github.therealmone.fireres.core.generator.firemode;
-
-import io.github.therealmone.fireres.core.pressure.ExcessPressureFactory;
-import io.github.therealmone.fireres.core.common.model.DoublePoint;
-import lombok.val;
-import org.junit.Test;
-
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-import static io.github.therealmone.fireres.core.TestUtils.*;
-
-public class ExcessPressureGeneratorTest {
-
-    @Test
-    public void generateTest() {
-        val props = defaultGenerationProperties();
-        val factory = new ExcessPressureFactory(props);
-
-        val pressure = factory.excessPressure();
-
-        val upperBound = IntStream.range(0, props.getGeneral().getTime())
-                .mapToObj(t -> new DoublePoint(t, props.getGeneral().getExcessPressure().getDelta()))
-                .collect(Collectors.toList());
-
-        val lowerBound = IntStream.range(0, props.getGeneral().getTime())
-                .mapToObj(t -> new DoublePoint(t, -props.getGeneral().getExcessPressure().getDelta()))
-                .collect(Collectors.toList());
-
-        assertFunctionNotHigher(pressure.getValue(), upperBound);
-        assertFunctionNotLower(pressure.getValue(), lowerBound);
-    }
-}
